# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: CR0SS.org API
  description: |
    Personal website API for habit tracking, AI chat, search, and content management.

    ## Authentication
    Most endpoints require authentication via the `X-Secret` header containing your secret token.
    Dashboard endpoints require the `X-Admin-Secret` header for enhanced security.

    ## Rate Limiting
    All endpoints are rate limited. Rate limit headers are included in responses:
    - `Retry-After`: Seconds until retry is allowed (when rate limited)

    Rate limits vary by endpoint:
    - AI Chat: 10 requests per 12 hours
    - Search: 100 requests per 60 seconds
    - Habits: 30 requests per 60 seconds
    - Dashboard: 30 requests per 60 seconds
    - Revalidation: 5 requests per 300 seconds (5 minutes)

  version: 1.0.0
  contact:
    name: CR0SS
    url: https://cr0ss.org
servers:
  - url: https://cr0ss.org/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
security:
  - SecretAuth: []
...
paths:
...
  /algolia/search:
    get:
      tags: [Search]
      summary: Search blog posts
      description: |
        Search through blog posts using Algolia.
        Supports both text search and click tracking.

        **Rate Limit:** 100 requests per 60 seconds
      security: []
      parameters:
        - name: q
          in: query
          description: Search query
          schema:
            type: string
          example: 'typescript'
        - name: objectID
          in: query
          description: Object ID for click tracking (when provided, tracks click instead of searching)
          schema:
            type: string
      responses:
        '200':
          description: Search results or click tracking confirmation
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      hits:
                        type: array
                        items:
                          $ref: '#/components/schemas/SearchHit'
                      queryID:
                        type: string
                        nullable: true
                  - type: object
                    properties:
                      success:
                        type: boolean
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    SecretAuth:
      type: apiKey
      in: header
      name: X-Secret
      description: Secret token for authentication
    AdminAuth:
      type: apiKey
      in: header
      name: X-Admin-Secret
      description: Admin secret token for dashboard endpoints
  schemas:
    SearchHit:
      type: object
      properties:
        objectID:
          type: string
        title:
          type: string
        summary:
          type: string
        author:
          type: string
        url:
          type: string
        image:
          type: string
          nullable: true
        categories:
          type: array
          items:
            type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          description: Additional error details
        timestamp:
          type: integer
          format: int64
  responses:
...
    RateLimitError:
      description: Rate limit exceeded
      headers:
        Retry-After:
          description: Seconds until retry is allowed
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: 'Too many requests'
            code: 'RATE_LIMIT_EXCEEDED'
            details:
              retryAfterSec: 300
              limit: 10
              windowHours: 12
            timestamp: 1706184000000
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: 'Internal server error'
            code: 'INTERNAL_ERROR'
            timestamp: 1706184000000